+++ tests/test_nuxeo.py
@@ -151,7 +151,10 @@ def test_file_out(tmp_path, server):
         check_lock += 1
 
     file_out = operation.execute(
-        file_out=tmp_path / "file_out", callback=callback, lock_path=lock_path, unlock_path=unlock_path
+        file_out=tmp_path / "file_out",
+        callback=callback,
+        lock_path=lock_path,
+        unlock_path=unlock_path,
     )
 
     try:
@@ -368,3 +371,32 @@ def test_unauthorized(server):
         assert text(e.value).startswith("Unauthorized(401)")
     finally:
         server.client.auth = auth
+
+
+def test_param_format(server, recwarn):
+    params = "stringnotadict"
+    with pytest.raises(HTTPError):
+        server.client.request("GET", "test", params=params)
+
+    assert len(recwarn) == 0
+
+    params = {"test.wrong.typo": "error"}
+    with pytest.warns(
+        DeprecationWarning,
+        match=r"'test.wrong.typo' param should not contain '_' nor '.'. Replace with '-' "
+        r"to get rid of that warning.",
+    ):
+        with pytest.raises(HTTPError):
+            server.client.request("GET", "test", params=params)
+
+
+def test_header_format(server):
+    headers = {"test.wrong.typo": "error"}
+
+    with pytest.warns(
+        DeprecationWarning,
+        match=r"'test.wrong.typo' header should not contain '_' nor '.'. Replace with '-' "
+        r"to get rid of that warning.",
+    ):
+        with pytest.raises(HTTPError):
+            server.client.request("GET", "test", headers=headers)